Fixed delete collab aPI

+++ emogo/apps/collaborator/views.py
@@ -137,8 +137,10 @@ class CollaboratorDeletionAPI(DestroyAPIView):
         :return: Hard Delete collaborator
         """
         collaborator = Collaborator.objects.get(id=kwargs.get('pk'))
-        noti = Notification.objects.filter(notification_type = 'collaborator_confirmation' , stream = collaborator.stream, from_user = self.request.user, to_user = User.objects.get(username = collaborator.phone_number) )
-        if noti.__len__() > 0 :
-            noti[0].delete()
+        collab_user = User.objects.filter(username = collaborator.phone_number)
+        if collab_user.__len__():
+            noti = Notification.objects.filter(notification_type = 'collaborator_confirmation' , stream = collaborator.stream, from_user = self.request.user, to_user = collab_user[0] )
+            if noti.__len__() > 0 :
+                noti[0].delete()
         collaborator.delete()
         return custom_render_response(status_code=status.HTTP_200_OK)
